src/pyproject.toml
[tool.poetry]
name = "py-content-safety"
version = "0.1.0"
description = ""
authors = ["Ricardo Cataldi <rcataldi@microsoft.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
azure-cognitiveservices-vision-contentmoderator = "^1.0.0"
azure-identity = "^1.19.0"
azure-ai-ml = "*"
azure-ai-contentsafety = "*"
python-dotenv = "^1.0.1"
azure-ai-inference = "*"
aiohttp = "^3.11.13"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
